refactor(go): separate middleware and router from main
