+++ go/svc/admin/pkg/controller/rest/router.go
@@ -1,17 +1,18 @@
-package main
+package rest
 
 import (
 	"net/http"
 
 	"github.com/go-chi/chi"
 
+	"github.com/ispec-inc/monorepo/go/pkg/middleware"
 	"github.com/ispec-inc/monorepo/go/pkg/presenter"
 	v1 "github.com/ispec-inc/monorepo/go/svc/admin/pkg/controller/rest/v1"
 )
 
 func NewRouter() http.Handler {
 	r := chi.NewRouter()
-	r = commonMiddleware(r)
+	r = middleware.Common(r)
 
 	r.Mount("/v1", v1.NewRouter())
